automatically backed up by learn

+++ lib/sql_queries.rb
@@ -30,7 +30,7 @@ def selects_the_titles_and_amount_over_goal_of_all_projects_that_have_met_their_
 end
 
 def selects_user_names_and_amounts_of_all_pledges_grouped_by_name_then_orders_them_by_the_amount_and_users_name
-"SELECT Users.name, CAST(SUM(Pledges.amount) as INT)
+"SELECT Users.name, CAST(SUM(Pledges.amount) as REAL)
   FROM Users
   LEFT OUTER JOIN Pledges
   ON Users.id = Pledges.user_id
@@ -39,7 +39,7 @@ def selects_user_names_and_amounts_of_all_pledges_grouped_by_name_then_orders_th
 end
 
 def selects_the_category_names_and_pledge_amounts_of_all_pledges_in_the_music_category
-"SELECT Projects.category, CAST(Pledges.amount as REAL)
+"SELECT Projects.category, CAST(Pledges.amount as INT)
 From Projects
 LEFT OUTER JOIN Pledges
 ON Projects.id = Pledges.project_id
